conflicting types in prototype

+++ in_class_code/09162022/pass_by_value.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 // function prototype declaration
-float double_my_number(int n);
+float double_my_number(float n);
 
 
 int main() {
